p2p: fix implicit conversion error in CAddrMan::Connected_()
as CAddress::nTime is declared in src/protocol.h as
uint32_t nTime{TIME_INIT};

Found by the addrman fuzzer with UBSan:

addrman.cpp:535:22: runtime error: implicit conversion
  from type 'int64_t' (aka 'long') of value 68719478016 (64-bit, signed)
  to type 'uint32_t' (aka 'unsigned int') changed the value to 1280 (32-bit, unsigned)
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior addrman.cpp:535:22

+++ src/addrman.cpp
@@ -532,7 +532,7 @@ void CAddrMan::Connected_(const CService& addr, int64_t nTime)
     // update info
     int64_t nUpdateInterval = 20 * 60;
     if (nTime - info.nTime > nUpdateInterval)
-        info.nTime = nTime;
+        info.nTime = static_cast<uint32_t>(nTime);
 }
 
 void CAddrMan::SetServices_(const CService& addr, ServiceFlags nServices)
